bots/shadowsafe/requirements.txt
python-telegram-bot>=21.0.0,<22.0.0
tomli; python_version<'3.11'
Pillow>=10.0.0
pikepdf>=8.5.0
oletools>=0.60
yara-python>=4.4.0
mutagen>=1.47.0
